# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ISRO ISTRAC SOC AI Platform",
    "description": "\n        AI-Driven Security Analytics Platform for ISRO ISTRAC Bengaluru.\n\n        ## Architecture\n        This platform ingests security logs from Elasticsearch, detects anomalies\n        using ML models (Isolation Forest, Autoencoder, LSTM), scores threats,\n        and provides natural language investigation via an SLM assistant.\n\n        ## Authentication\n        All endpoints require Bearer token authentication (see /api/auth/login).\n\n        ## Rate Limits\n        - /api/slm/chat: 60 requests/minute\n        - /api/alerts: 300 requests/minute\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/incidents": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_api_incidents_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "attack_stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Attack Stage"
            }
          },
          {
            "name": "threat_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Threat Level"
            }
          },
          {
            "name": "host_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Host Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Incidents Api Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/stats": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "Get Incident Stats",
        "operationId": "get_incident_stats_api_incidents_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/incidents/{incident_id}": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "Get Incident Detail",
        "operationId": "get_incident_detail_api_incidents__incident_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/status": {
      "patch": {
        "tags": [
          "Incidents"
        ],
        "summary": "Update Incident Status",
        "operationId": "update_incident_status_api_incidents__incident_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/escalate": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Escalate Incident",
        "operationId": "escalate_incident_api_incidents__incident_id__escalate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentEscalation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/investigate": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "Investigate Incident",
        "operationId": "investigate_incident_api_incidents__incident_id__investigate_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/generate-report": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Trigger Report Generation",
        "operationId": "trigger_report_generation_api_incidents__incident_id__generate_report_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/report": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "Fetch Incident Report",
        "operationId": "fetch_incident_report_api_incidents__incident_id__report_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentDetailResponse": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "entity_key": {
            "type": "string",
            "title": "Entity Key"
          },
          "host_id": {
            "type": "string",
            "title": "Host Id"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "alert_count": {
            "type": "integer",
            "title": "Alert Count"
          },
          "log_types_involved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Log Types Involved"
          },
          "max_threat_score": {
            "type": "number",
            "title": "Max Threat Score"
          },
          "incident_threat_score": {
            "type": "number",
            "title": "Incident Threat Score"
          },
          "threat_level": {
            "type": "string",
            "title": "Threat Level"
          },
          "mitre_tactics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mitre Tactics"
          },
          "mitre_techniques": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mitre Techniques"
          },
          "attack_stage": {
            "type": "string",
            "title": "Attack Stage"
          },
          "is_multi_stage": {
            "type": "boolean",
            "title": "Is Multi Stage"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "matched_patterns": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Patterns"
          },
          "alerts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Alerts"
          },
          "timeline": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Timeline"
          },
          "attack_chain": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Attack Chain"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "entity_key",
          "host_id",
          "started_at",
          "last_seen",
          "duration_seconds",
          "alert_count",
          "log_types_involved",
          "max_threat_score",
          "incident_threat_score",
          "threat_level",
          "mitre_tactics",
          "mitre_techniques",
          "attack_stage",
          "is_multi_stage",
          "status",
          "created_at",
          "alerts",
          "timeline",
          "attack_chain"
        ],
        "title": "IncidentDetailResponse"
      },
      "IncidentEscalation": {
        "properties": {
          "escalated_to": {
            "type": "string",
            "title": "Escalated To"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "escalated_to",
          "reason"
        ],
        "title": "IncidentEscalation"
      },
      "IncidentStatsResponse": {
        "properties": {
          "total_active": {
            "type": "integer",
            "title": "Total Active"
          },
          "multi_stage_count": {
            "type": "integer",
            "title": "Multi Stage Count"
          },
          "by_attack_stage": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Attack Stage"
          },
          "by_threat_level": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Threat Level"
          },
          "avg_duration_minutes": {
            "type": "number",
            "title": "Avg Duration Minutes"
          },
          "top_targeted_hosts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top Targeted Hosts"
          }
        },
        "type": "object",
        "required": [
          "total_active",
          "multi_stage_count",
          "by_attack_stage",
          "by_threat_level",
          "avg_duration_minutes",
          "top_targeted_hosts"
        ],
        "title": "IncidentStatsResponse"
      },
      "IncidentStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "IncidentStatusUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}